+++ src/lib/core/utility/find.ts
@@ -1,4 +1,12 @@
-import { BaseType, Selection } from 'd3-selection';
+import { BaseType, Selection, ValueFn } from 'd3-selection';
+
+export function findByFilter<GElement extends BaseType, Datum = unknown>(
+  container: Selection,
+  selector: string,
+  filter: ValueFn<GElement, Datum, boolean>
+): Selection<GElement, Datum> {
+  return container.selectAll<GElement, Datum>(selector).filter(filter);
+}
 
 export function findByDataProperty<GElement extends BaseType, Datum>(
   container: Selection,
@@ -6,15 +14,15 @@ export function findByDataProperty<GElement extends BaseType, Datum>(
   property: keyof Datum,
   value: any
 ): Selection<GElement, Datum> {
-  return container.selectAll<GElement, Datum>(selector).filter((d) => d[property] === value);
+  return findByFilter<GElement, Datum>(container, selector, (d) => d[property] === value);
 }
 
 export function findByKey<GElement extends BaseType, Datum extends { key: string | number }>(
   container: Selection,
   selector: string,
-  value: any
+  key: string | number
 ): Selection<GElement, Datum> {
-  return container.selectAll<GElement, Datum>(selector).filter((d) => d['key'] === value);
+  return findByDataProperty<GElement, Datum>(container, selector, 'key', key);
 }
 
 export function findByIndex<GElement extends BaseType, Datum = unknown>(
@@ -22,5 +30,5 @@ export function findByIndex<GElement extends BaseType, Datum = unknown>(
   selector: string,
   index: number
 ): Selection<GElement, Datum> {
-  return container.selectAll<GElement, Datum>(selector).filter((d, i) => i === index);
+  return findByFilter<GElement, Datum>(container, selector, (d, i) => i === index);
 }